DocuShield-AI/requirements.txt
flask
werkzeug
PyPDF2
python-docx
python-pptx
openpyxl
odfpy
